requirements.txt
lxml>=3.6.4
pyyaml>=3.12